removed extra lines and changed the file name
