+++ C++/13-Bit Manipulation/14-xor_of_array_elements.cpp
@@ -1,0 +1,24 @@
+	#include<bits/stdc++.h>
+	using namespace std;
+	int main()
+	{
+		int T;
+		cin>>T;
+		while(T--)
+		{
+			int N;
+			cin>>N;
+			int arr[N];
+			for(int i=0;i<N;i++)
+			{
+				cin>>arr[i];
+			}
+			int x=arr[0];
+			for(int i=1;i<N;i++)
+			{
+				x=x | arr[i];
+			}
+			cout<<x<<endl;
+		}
+		return 0;
+	}
