+++ C++/13-Bit Manipulation/2-check_ith_bit_is_set_or_not.cpp
@@ -1,0 +1,21 @@
+	#include<bits/stdc++.h>
+	using namespace std;
+	int main()
+	{
+		int t;
+		cin>>t;
+		while(t--)
+		{
+			int n,i;
+			cin>>n>>i;
+			//F=00000001  left shift F by i times and perform AND operation with N, if the result is 0 the the ith bit is not set else the ith bit is set
+			int f=1;
+			f=f<<i;
+			int result=n & f;
+			if(result==0)
+			cout<<"ith bit is not set\n";
+			else
+			cout<<"ith bit is set\n";
+		}
+		return 0;
+	}
