--- C++/13-Bit Manipulation/3 bitwise left shift.cpp
@@ -1,51 +1,0 @@
-	//Diameter of a Tree or length of the longest path
-
-	#include<bits/stdc++.h>
-	//#define<windows.h>
-	//prime fctorization  
-	//#include<ext.pb_ds/assoc_container.hpp>
-	//using namespace __gnu_pbds;
-	using namespace std;
-
-	#define ff    	   		first
-	#define ss    			second
-	#define ll      		long long
-	#define pb    			push_back
-	#define					REP(i,n) for (int i=1;i<=n;i++)
-	#define mp    			make_pair
-	#define pii   			pair<int,int>
-	#define vi	  			vector<int>
-	#define vii 			vector<pii>
-	#define					endl  '\n'
-	#define mii   			map<int,int>
-	#define pqb   			priority_queue<int>
-	#define pqs				priority_queue<int,vi,greater<int> >
-	#define setbits(x)		__builtin_popcountll(x)
-	#define zrobits(x)		__builtin_ctzll(x)
-	#define mod             1000000007
-	#define INF				1000000000
-	#define ps(x,y)			fixed<<setprecision(y)<<x
-	#define mk(arr,n,type)	type *arr=new type[n];
-	#define w(x)			int x; cin>>x; while(x--)
-	const double PI=3.141592653589793238460;
-	typedef std::complex<double> Complex;
-
-
-	int main()
-	{
-		int t;
-		cin>>t;
-		while(t--)
-		{
-			int n,i;
-			
-			cin>>n>>i;
-			
-			//Left shift n by i bits i.e mutiplying n by 2^i
-			
-			cout<<n<<"<<"<<i<<" = "<<(n<<i)<<endl;
-		}
-		
-		return 0;
-	}
-
