--- C++/13-Bit Manipulation/All pair sum Xor.cpp
@@ -1,83 +1,0 @@
-
-
-#include<bits/stdc++.h>
-//#define<windows.h>
-//prime fctorization  
-//#include<ext.pb_ds/assoc_container.hpp>
-//using namespace __gnu_pbds;
-using namespace std;
-
-#define ff    	   		first
-#define ss    			second
-#define ll      		long long
-#define pb    			push_back
-#define					REP(i,n) for (int i=1;i<=n;i++)
-#define mp    			make_pair
-#define pii   			pair<int,int>
-#define vi	  			vector<int>
-#define vii 			vector<pii>
-#define					endl  '\n'
-#define mii   			map<int,int>
-#define pqb   			priority_queue<int>
-#define pqs				priority_queue<int,vi,greater<int> >
-#define setbits(x)		__builtin_popcountll(x)
-#define zrobits(x)		__builtin_ctzll(x)
-#define mod             1000000007
-#define INF				1000000000
-#define ps(x,y)			fixed<<setprecision(y)<<x
-#define mk(arr,n,type)	type *arr=new type[n];
-#define w(x)			int x; cin>>x; while(x--)
-const double PI=3.141592653589793238460;
-typedef std::complex<double> Complex;
-vi vect;
-
-/*Question: Given an array of n integers, find the sum of xor of all pairs of numbers in the array.
-
-Examples :
-
-Input : arr[] = {7, 3, 5}
-Output : 12
-7 ^ 3 = 4
-3 ^ 5 = 6
-7 ^ 5 = 2
-Sum = 4 + 6 + 2 
-    = 12
-
-Input : arr[] = {5, 9, 7, 6}
-Output : 47
-5 ^ 9 = 12
-9 ^ 7 = 14
-7 ^ 6 = 1
-5 ^ 7 = 2
-5 ^ 6 = 3
-9 ^ 6 = 15
-Sum = 12 + 14 + 1 + 2 + 3 + 15
-    = 47
-*/
-int32_t main()
-{
-
-	//ios_base::sync_with_stdio(0); cin.tie(0); cout.tie(0);
-	
-       
-      w(t)
-      {
-      	
-	   int n;
-	   cin>>n;
-	   int res=0;
-	   for(int i=0;i<n;i++)
-	   {
-	   	 int m;
-	   	 cin>>m;
-	   	 res^=2*m;
-	   }
-	   
-	   cout<<res<<endl;
-      	
-      	
-	  }
-	
-	return 0;
-}
-
